ButtonClick now has it's own Listener class
